Conflicts:
	backend/app/controllers/organizations.js
	backend/app/controllers/products.js
	backend/app/controllers/projects.js
	backend/app/controllers/survey_answers.js
	backend/app/controllers/surveys.js
	backend/app/controllers/tasks.js
	backend/app/controllers/users.js
	backend/app/controllers/workflows.js
	backend/app/models/product_uoa.js
	backend/app/router.js

+++ client/app/greyscale.tables/scripts/services/greyscale-project-products.js
@@ -45,6 +45,7 @@ angular.module('greyscale.tables')
             show: true,
             sortable: 'surveyId',
             dataFormat: 'option',
+            cellTemplate: '<span ng-if="option.id">{{option.title}} <small>(<span ng-show="option.isDraft" translate="SURVEYS.IS_DRAFT"></span><span ng-show="!option.isDraft" translate="SURVEYS.IS_COMPLETE"></span>)</small></span>',
             //dataRequired: true,
             dataSet: {
                 getData: _getSurveys,
@@ -61,6 +62,10 @@ angular.module('greyscale.tables')
             sortable: 'workflow.name',
             title: tns + 'WORKFLOW',
             show: true,
+            cellTemplate: '{{cell}}<span ng-if="!cell" class="action" translate="' + tns + 'CREATE_WORKFLOW"></span>',
+            link: {
+                handler: _editProductWorkflow
+            },
             dataHide: true
         }, {
             field: 'status',
@@ -90,10 +95,6 @@ angular.module('greyscale.tables')
                 class: 'info',
                 handler: _editProductUoas
             }, {
-                title: tns + 'WORKFLOW',
-                class: 'info',
-                handler: _editProductWorkflow
-            }, {
                 title: tns + 'TASKS',
                 class: 'info',
                 handler: _editProductTasks
@@ -102,14 +103,12 @@ angular.module('greyscale.tables')
             show: true,
             dataFormat: 'action',
             actions: [{
-                title: '',
                 icon: 'fa-pencil',
-                class: 'info',
+                tooltip: 'COMMON.EDIT',
                 handler: _editProduct
             }, {
-                title: '',
                 icon: 'fa-trash',
-                class: 'danger',
+                tooltip: 'COMMON.DELETE',
                 handler: _removeProduct
             }]
         }];
@@ -126,7 +125,6 @@ angular.module('greyscale.tables')
             formWarning: _getFormWarning,
             pageLength: 10,
             add: {
-                title: 'COMMON.CREATE',
                 handler: _editProduct
             }
         };
@@ -186,11 +184,18 @@ angular.module('greyscale.tables')
         }
 
         function _removeProduct(product) {
-            greyscaleProductApi.delete(product.id)
-                .then(_reload)
-                .catch(function (err) {
-                    inform.add('Product delete error: ' + err);
-                });
+            greyscaleModalsSrv.confirm({
+                message: tns + 'DELETE_CONFIRM',
+                product: product,
+                okType: 'danger',
+                okText: 'COMMON.DELETE'
+            }).then(function () {
+                greyscaleProductApi.delete(product.id)
+                    .then(_reload)
+                    .catch(function (err) {
+                        inform.add('Product delete error: ' + err);
+                    });
+            });
         }
 
         function _reload() {
