+++ client/app/greyscale.tables/scripts/services/greyscale-project-surveys.js
@@ -20,24 +20,26 @@ angular.module('greyscale.tables')
             dataRequired: false,
             dataFormat: 'text'
         }, {
+            field: 'isDraft',
+            title: tns + 'STATUS',
+            show: true,
+            cellTemplate: '<span ng-if="cell" class="text-warning" translate="SURVEYS.IS_DRAFT"></span><span ng-if="!cell" class="text-success" translate="SURVEYS.IS_COMPLETE"></span>'
+        }, {
             field: '',
             title: '',
             show: true,
             dataFormat: 'action',
             actions: [{
-                title: '',
                 icon: 'fa-eye',
-                class: 'info',
+                tooltip: 'COMMON.VIEW',
                 handler: _viewSurvey
             }, {
-                title: '',
                 icon: 'fa-pencil',
-                class: 'info',
+                tooltip: 'COMMON.EDIT',
                 handler: _editSurvey
             }, {
-                title: '',
                 icon: 'fa-trash',
-                class: 'danger',
+                tooltip: 'COMMON.DELETE',
                 handler: _deleteSurvey
             }]
         }];
@@ -51,7 +53,6 @@ angular.module('greyscale.tables')
             dataFilter: {},
             formTitle: tns + 'ITEM',
             add: {
-                title: 'COMMON.CREATE',
                 handler: _editSurvey
             }
         };
@@ -87,14 +88,24 @@ angular.module('greyscale.tables')
         }
 
         function _deleteSurvey(_survey) {
-            greyscaleSurveyApi.delete(_survey).then(_reload).catch(function (err) {
-                inform.add('Survey delete error: ' + err);
+            greyscaleModalsSrv.confirm({
+                message: tns + 'DELETE_CONFIRM',
+                survey: _survey,
+                okType: 'danger',
+                okText: 'COMMON.DELETE'
+            }).then(function () {
+                greyscaleSurveyApi.delete(_survey).then(_reload).catch(function (err) {
+                    inform.add('Survey delete error: ' + err);
+                });
             });
         }
 
         function _viewSurvey(_survey) {
-            window.location = $location.protocol() + '://' + $location.host() + ':' + $location.port() +
-                '/interviewRenderer/?sureveyId=' + _survey.id;
+            //window.location = $location.protocol() + '://' + $location.host() + ':' + $location.port() +
+            //    '/interviewRenderer/?sureveyId=' + _survey.id;
+            $state.go('survey', {
+                surveyId: _survey.id
+            });
         }
 
         return _table;
